requirements.txt
fastapi
uvicorn
rembg
opencv-python-headless
numpy
pillow
scikit-image
trimesh
python-multipart
pyglet
pyopengl
onnxruntime
